Move variables and functions into new files
	Add player.h
	Add player.cpp
	The game is having an issue that makes it crashed after pressing
	goButton, need to investigate more
